rename: components 오타 수정
