+++ src/pages/Home/Home.jsx
@@ -1,7 +1,7 @@
 import * as S from "./Home.style.js";
 import Loading from "../Loading/Loading.jsx";
-import Navbar from "../../componenets/Navbar/Navbar.jsx";
-import LectureList from "../../componenets/LectureList/LectureList.jsx";
+import Navbar from "../../components/Navbar/Navbar.jsx";
+import LectureList from "../../components/LectureList/LectureList.jsx";
 import { useEffect, useState } from "react";
 import useFetch from "../../hooks/useFetch.js";
 import { ENDPOINTS } from "../../api/endpoints.js";
